requirements.txt
pathlib
google-cloud-storage
